+++ Assets/LightweightAsset.asset
@@ -14,7 +14,7 @@ MonoBehaviour:
   k_AssetVersion: 3
   m_MaxPixelLights: 1
   m_SupportsVertexLight: 0
-  m_RequireDepthTexture: 0
+  m_RequireDepthTexture: 1
   m_RequireSoftParticles: 0
   m_RequireOpaqueTexture: 1
   m_OpaqueDownsampling: 1
@@ -37,4 +37,11 @@ MonoBehaviour:
   m_KeepLocalShadowVariants: 1
   m_KeepSoftShadowVariants: 1
   m_ResourcesAsset: {fileID: 11400000, guid: aac5a08c32552a14c89394b703f1978a, type: 2}
+  m_SavedXRConfig:
+    renderScale: 1
+    viewportScale: 1
+    useOcclusionMesh: 1
+    occlusionMaskScale: 1
+    showDeviceView: 1
+    gameViewRenderMode: 3
   m_ShadowType: 2
