+++ Assets/Scripts/Rendering.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 6748847e6642c4fd3888155d0874af7d
-NativeFormatImporter:
+guid: de0358983d1724e078d3f85e29d3d0a8
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
