+++ Assets/Scripts/Rendering/WaterFXPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c61a0c521779544ee8bf7890846c4d21
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
