+++ Assets/Shaders/PackedPBR.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 169ccfffe27834dd89f43a5f0edc67ec
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
