+++ Assets/Shaders/PackedPBR/InputSurfacePackedDialectric.hlsl.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9206d98c6f87d47f78c5c12e4095ceb3
+guid: 6ff1bcd73ddf944269e34f80b93fc017
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
