+++ Assets/Shaders/PackedPBR/LightweightPackedDialectric.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b2f73566a373e49e18ecd37f9e9f2386
+guid: dc06203bc9dae41408e28313d7eab8ed
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
