+++ Assets/scenes/Testing/PackedPBR shader.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77a2fc799d4174f6296bedcb42fe5097
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
