+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.2.0b8
+m_EditorVersion: 2018.2.0b11
